From:    Markus Gyger <mgyger@itr.ch> here are some minor cosmetic changes for squid-1.1.beta6. BTW: configure doesn't reflect configure.in correctly right now.

+++ src/client_side.c
@@ -305,7 +305,7 @@ proxyAuthenticate(char *headers)
 
 		passwords = xmalloc((size_t) buf.st_size + 2);
 		f = fopen(Config.proxyAuthFile, "r");
-		fread(passwords, buf.st_size, 1, f);
+		fread(passwords, (size_t) buf.st_size, 1, f);
 		*(passwords + buf.st_size) = '\0';
 		strcat(passwords, "\n");
 		fclose(f);
